move raw_data.utils -> jitclasses
